Fixes from feedback 7
